feat: 환율정보 뷰,뷰모델로 분리
